refactor: web, console 폴더 분리
